+++ just_audio/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.10.2
 
+* Add Player.playerEvent.
 * Fix playing/playbackEvent emission order.
 * Fix rxdart errors on lower bound rxdart 0.26.0.
 * Fix MissingPluginException in dispose on iOS.